Eucalyptus dataset

This dataset represents a clonal eucalyptus trial. It contains the mean annual increment (MAI) of 100 clones. The trial was laid out in randomized complete blocks design, with 13 replicates and one tree per plot. This trial did not have contiguous blocks: the first six blocks were in area 1 and the last seven were in area 2. The area column contains this information. The dataset also has information about the position of each tree in the grid (row and col). In the row and column directions, the plants are spaced 2 m (dist_row) and 3 m (dist_col) apart, respectively. Data were collected at 3 and 6 years after the trial implementation (information contained in the column age).

euca

Format

euca

A data frame with 4200 rows and 10 columns:

area

2 areas

age

2 ages

tree

1144 trees

block

13 blocks

clone

100 clones

dist_row

2 meters between rows

dist_col

3 meters between columns

row

26 rows

col

44 columns

MAI

Mean annual increment of wood volume (m3 ha-1 year-1)