gencomp: Fit Spatial-Genetic Competition Models

Fit spatial-genetic competition models for assessing the competition capacity of genotypes. Contained are functions used to build a competition matrix based on different criteria, fit the spatial-genetic competition model, and extract and visualize the results.

See also

Author

Maintainer: Saulo F. S. Chaves saulo.chaves@ufv.br (ORCID)

Authors:

Kaio Olimpio G. Dias kaio.o.dias@ufv.br (ORCID) [copyright holder]

Filipe M. Ferreira ferreira.fmanoel@gmail.com (ORCID)

Getúlio C. Ferreira getuliocaifer@gmail.com (ORCID)

Salvador A. Gezan forestats.sg@gmail.com (ORCID)