Print an object of class comprepfor

Print a comprepfor object in the R console.

# S3 method for class 'comprepfor'
print(x, ..., category = "matrix", age = "all")

Arguments

x

An object of class comprepfor

...

Currently not used.

category

A string indicating which object to print. Options are "all" for printing all objects, "data" for printing the data that will be used in the model, "matrix" for printing the competition matrix, "check" for printing the neigh_check dataframe, and "CIF" for printing the competition intensity factor.

age

A string indicating if objects should be printed per age. Options are "all" for printing all ages, or the name of the specific age. Defaults to "all", which also serves when data has a single age.

See also