Print an object of class comresp

Print a comresp object in the R console.

# S3 method for class 'comresp'
print(x, ..., category = "blup.main", age = "all")

Arguments

x

An object of class comresp.

...

Currently not used.

category

A string indicating which object to print. Options are "all" for printing all objects, "summar" for printing the variance components and the likelihood ratio test (if lrt = TRUE in asr or asr_ma), "blup.main" (Default) for printing the DGE, IGE and TGV (the main effects if a multi-age model was fitted), and "blup.within" for printing the DGE, IGE and TGV within ages (if a multi-age model was fitted).

age

If category = 'blup.within', a string indicating if this object should be printed per age. Options are "all" for printing all ages, or the name of a specific age. Defaults to "all".

See also

Examples

# \donttest{
 library(gencomp)
 comp_mat = prepfor(data = euca, gen = 'clone', area = 'area',
                   plt = 'tree', age = 'age', row = 'row', col = 'col',
                   dist.col = 3, dist.row = 2, trait = 'MAI', method = 'SK',
                   n.dec = 3, verbose = FALSE, effs = c("block"))
 model = asr_ma(prep.out = comp_mat,
                fixed = MAI~ age, 
                random = ~ block:age, 
                lrtest = TRUE, 
                spatial = TRUE, 
                cor = TRUE, 
                maxit = 20)
#> ASReml Version 4.2 21/01/2025 18:22:00
#>           LogLik        Sigma2     DF     wall
#>  1     -5695.638           1.0   1645   18:22:01  (  4 restrained)
#>  2     -5632.490           1.0   1645   18:22:01  (  5 restrained)
#>  3     -5562.100           1.0   1645   18:22:02  (  3 restrained)
#>  4     -5515.785           1.0   1645   18:22:02  (  2 restrained)
#>  5     -5489.058           1.0   1645   18:22:03  (  1 restrained)
#>  6     -5461.928           1.0   1645   18:22:03  (  1 restrained)
#>  7     -5447.018           1.0   1645   18:22:03
#>  8     -5442.411           1.0   1645   18:22:04
#>  9     -5441.232           1.0   1645   18:22:04
#> 10     -5441.178           1.0   1645   18:22:04
#> 11     -5441.176           1.0   1645   18:22:05
#> ====> Starting likelihood ratio tests
#> ASReml Version 4.2 21/01/2025 18:22:05
#>           LogLik        Sigma2     DF     wall
#>  1     -5567.226           1.0   1645   18:22:06  (  1 restrained)
#>  2     -5502.968           1.0   1645   18:22:06  (  3 restrained)
#>  3     -5479.250           1.0   1645   18:22:06  (  1 restrained)
#>  4     -5460.279           1.0   1645   18:22:07  (  1 restrained)
#>  5     -5453.254           1.0   1645   18:22:07  (  1 restrained)
#>  6     -5450.791           1.0   1645   18:22:07  (  1 restrained)
#>  7     -5449.892           1.0   1645   18:22:07
#>  8     -5449.753           1.0   1645   18:22:08
#>  9     -5449.751           1.0   1645   18:22:08
#> ASReml Version 4.2 21/01/2025 18:22:08
#>           LogLik        Sigma2     DF     wall
#>  1     -5596.539           1.0   1645   18:22:09  (  1 restrained)
#>  2     -5528.593           1.0   1645   18:22:09  (  3 restrained)
#>  3     -5496.563           1.0   1645   18:22:09  (  1 restrained)
#>  4     -5469.550           1.0   1645   18:22:09  (  1 restrained)
#>  5     -5457.877           1.0   1645   18:22:09  (  1 restrained)
#>  6     -5453.135           1.0   1645   18:22:10  (  1 restrained)
#>  7     -5451.147           1.0   1645   18:22:10
#>  8     -5450.770           1.0   1645   18:22:10
#>  9     -5450.762           1.0   1645   18:22:10
#> ASReml Version 4.2 21/01/2025 18:22:10
#>           LogLik        Sigma2     DF     wall
#>  1     -5538.532           1.0   1645   18:22:11
#>  2     -5501.241           1.0   1645   18:22:11  (  2 restrained)
#>  3     -5479.725           1.0   1645   18:22:11  (  1 restrained)
#>  4     -5466.675           1.0   1645   18:22:11
#>  5     -5455.535           1.0   1645   18:22:11
#>  6     -5450.825           1.0   1645   18:22:11
#>  7     -5449.790           1.0   1645   18:22:11
#>  8     -5449.751           1.0   1645   18:22:11
#>  9     -5449.751           1.0   1645   18:22:12
#> ASReml Version 4.2 21/01/2025 18:22:12
#>           LogLik        Sigma2     DF     wall
#>  1     -5683.330           1.0   1645   18:22:12  (  1 restrained)
#>  2     -5604.604           1.0   1645   18:22:12  (  2 restrained)
#>  3     -5565.312           1.0   1645   18:22:13  (  2 restrained)
#>  4     -5541.890           1.0   1645   18:22:13  (  2 restrained)
#>  5     -5527.033           1.0   1645   18:22:13  (  1 restrained)
#>  6     -5512.722           1.0   1645   18:22:13  (  1 restrained)
#>  7     -5506.215           1.0   1645   18:22:13
#>  8     -5504.876           1.0   1645   18:22:13
#>  9     -5504.665           1.0   1645   18:22:14
#> 10     -5504.661           1.0   1645   18:22:14
#> ASReml Version 4.2 21/01/2025 18:22:14
#>           LogLik        Sigma2     DF     wall
#>  1     -5573.612           1.0   1645   18:22:14
#>  2     -5532.774           1.0   1645   18:22:14  (  2 restrained)
#>  3     -5509.337           1.0   1645   18:22:14
#>  4     -5489.122           1.0   1645   18:22:15
#>  5     -5480.358           1.0   1645   18:22:15
#>  6     -5478.228           1.0   1645   18:22:15
#>  7     -5478.111           1.0   1645   18:22:15
#>  8     -5478.108           1.0   1645   18:22:15
#> ====> LRT results:
#>    effect  LR-statistic    Pr(Chisq)
#> 1     DGE  1.098212e+02 0.000000e+00
#> 2     IGE  5.671546e+01 2.520206e-14
#> 3 DGE:age  2.022865e+00 7.747327e-02
#> 4 IGE:age -9.514352e-05 5.000000e-01
             
 results = resp(prep.out = comp_mat, model = model, weight.tgv = FALSE, sd.class = 1)
 
 print(results)
#> ===> DGE, IGE and TGV
#>       clone         DGE    se.DGE   rel.DGE          IGE   se.IGE   rel.IGE
#>      <fctr>       <num>     <num>     <num>        <num>    <num>     <num>
#>   1:   C099  24.6155934  3.577760 0.9106661  -0.79919798 2.460610 0.7061682
#>   2:   C087  13.2013688  4.119508 0.8815638   2.19377705 2.477019 0.7022363
#>   3:   C083  22.7097354  3.620486 0.9085197  -3.44198318 2.510407 0.6941551
#>   4:   C028  26.1498864  3.475277 0.9157107  -6.02446200 2.636438 0.6626753
#>   5:   C066   8.5283065  3.635331 0.9077680   1.33525136 2.334376 0.7355431
#>   6:   C092  28.1174949  3.666496 0.9061798  -7.11882264 2.645961 0.6602338
#>   7:   C025   0.3828506  3.581798 0.9104644   4.71115499 2.388269 0.7231915
#>   8:   C073  14.4757695  3.552133 0.9119413  -1.34904442 2.389080 0.7230034
#>   9:   C023  11.1799264  4.429750 0.8630531  -0.07818138 2.651615 0.6587803
#>  10:   C096   8.4122922  3.420055 0.9183681   0.97170205 2.498207 0.6971205
#>  11:   C033   0.9581647  3.997578 0.8884710   3.81976471 2.434505 0.7123699
#>  12:   C019  -1.7493381  5.512207 0.7879468   4.87424467 2.372559 0.7268210
#>  13:   C075  34.7749389  3.446719 0.9170903 -10.86980030 2.542849 0.6861990
#>  14:   C093  26.9481426  3.548553 0.9121187  -8.02511580 2.543072 0.6861440
#>  15:   C085   7.1142751  3.662342 0.9063923   0.27182076 2.525885 0.6903720
#>  16:   C076 -14.0265744 10.313925 0.2575935   9.12471633 2.237582 0.7570198
#>  17:   C046   3.6440155  3.469133 0.9160084   1.50515350 2.546888 0.6852015
#>  18:   C011   8.8306228  3.819568 0.8981825  -1.15897999 2.523569 0.6909396
#>  19:   C039 -12.0444837  4.058322 0.8850559   7.65086450 2.476385 0.7023886
#>  20:   C082  -3.6719377  3.551550 0.9119702   4.03958734 2.468875 0.7041910
#>  21:   C080   1.3295020  3.453747 0.9167518   1.81315897 2.522621 0.6911717
#>  22:   C008   0.6641248  3.423279 0.9182141   1.96879549 2.453173 0.7079417
#>  23:   C097  16.1257274  3.620008 0.9085439  -4.66396455 2.530240 0.6893034
#>  24:   C089  10.9336620  3.433417 0.9177290  -2.59197392 2.515964 0.6927996
#>  25:   C004  -8.1110852  4.080600 0.8837905   5.56384109 2.424791 0.7146605
#>  26:   C037  -8.8753419 10.513700 0.2285551   5.77368180 2.601459 0.6715667
#>  27:   C016 -10.5108365  6.807094 0.6766169   6.42060688 2.435875 0.7120460
#>  28:   C054   4.6542420  3.725631 0.9031291  -0.17742466 2.437575 0.7116439
#>  29:   C022  -3.3588666  3.811756 0.8985986   3.14989920 2.659521 0.6567426
#>  30:   C020   1.0856965  4.067402 0.8845409   1.12806025 2.426654 0.7142221
#>  31:   C038  12.7536946  3.685042 0.9052283  -3.91118102 2.600725 0.6717520
#>  32:   C029   9.1263106  3.929047 0.8922622  -2.37897041 2.632571 0.6636640
#>  33:   C048  -7.2699648  4.140305 0.8803650   4.62928876 2.518965 0.6920662
#>  34:   C030  -9.8780638  4.650783 0.8490455   5.74248980 2.312698 0.7404322
#>  35:   C091  16.3893815  3.722072 0.9033141  -5.52886191 2.467639 0.7044872
#>  36:   C042 -10.3757947  4.306107 0.8705913   5.70613591 2.532117 0.6888422
#>  37:   C094  12.7981424  3.556092 0.9117449  -4.34227862 2.568570 0.6798186
#>  38:   C012   1.0520764  3.961655 0.8904665   0.65555788 2.662183 0.6560551
#>  39:   C006   5.7488966  3.660083 0.9065077  -1.40256034 2.497139 0.6973793
#>  40:   C027  17.7637866  4.127863 0.8810829  -6.57678701 2.661813 0.6561507
#>  41:   C058  -4.4052486  4.538619 0.8562389   2.86180659 2.573175 0.6786698
#>  42:   C013  -4.4145988  3.830370 0.8976058   2.82351806 2.548220 0.6848720
#>  43:   C014   0.6429174  3.439070 0.9174578   0.61302333 2.495953 0.6976667
#>  44:   C055  20.3115351  3.588649 0.9101215  -7.88398321 2.639078 0.6619993
#>  45:   C098   8.9223245  3.825352 0.8978739  -3.02235243 2.395537 0.7215039
#>  46:   C010   4.4739674  3.521331 0.9134619  -1.11738481 2.609469 0.6695412
#>  47:   C057   9.2017902  3.533404 0.9128674  -3.33755044 2.549294 0.6846063
#>  48:   C069  -8.8977865  4.105342 0.8823769   4.34254687 2.798305 0.6199827
#>  49:   C078   0.2608237  3.607569 0.9091713   0.33712395 2.587089 0.6751851
#>  50:   C049   1.1412872  3.986351 0.8890966  -0.14664503 2.523758 0.6908933
#>  51:   C047  -0.9559314 10.559552 0.2218116   0.62186268 2.678899 0.6517222
#>  52:   C045  -7.6352539  4.138562 0.8804656   3.43113860 2.492540 0.6984930
#>  53:   C062   1.6158885  3.537829 0.9126491  -0.59025523 2.393447 0.7219899
#>  54:   C095 -13.8146351  4.668129 0.8479174   5.98575328 2.480467 0.7014068
#>  55:   C001   4.5943942  3.481016 0.9154321  -2.09570546 2.585117 0.6756802
#>  56:   C050   0.6659203 10.396883 0.2456028  -0.43320157 2.394549 0.7217338
#>  57:   C079 -11.7117120  4.362077 0.8672054   4.74144669 2.423588 0.7149436
#>  58:   C032  -3.4555615  3.835731 0.8973190   1.15480877 2.588327 0.6748742
#>  59:   C056   2.2619037  4.020630 0.8871811  -1.55786473 2.413082 0.7174098
#>  60:   C084   4.8118781  3.659456 0.9065398  -2.73563382 2.465959 0.7048895
#>  61:   C067   3.8597666  3.407732 0.9189553  -2.35914171 2.472562 0.7033068
#>  62:   C070   1.1098787  3.929629 0.8922303  -1.21558453 2.475821 0.7025242
#>  63:   C088   8.8995446  3.359176 0.9212484  -5.01457255 2.403720 0.7195980
#>  64:   C065  -9.7579765  3.937199 0.8918146   2.91772336 2.591672 0.6740334
#>  65:   C034  -6.1914023  4.461427 0.8610875   1.36788737 2.450179 0.7086542
#>  66:   C043 -12.9939921  6.487007 0.7063145   4.25653639 2.292927 0.7448513
#>  67:   C074   0.9348632  3.524303 0.9133157  -1.82321355 2.369375 0.7275539
#>  68:   C009   9.0797815  3.782424 0.9001532  -5.57642663 2.658715 0.6569505
#>  69:   C064  -4.3933717  3.552215 0.9119372   0.04741177 2.603557 0.6710367
#>  70:   C077 -10.4876080  3.957612 0.8906899   2.61736719 2.575751 0.6780259
#>  71:   C072  -0.3063273  3.349451 0.9217037  -1.75013945 2.423955 0.7148573
#>  72:   C005 -13.7764233  3.817664 0.8982840   3.95794324 2.507961 0.6947506
#>  73:   C068  -4.0034897  3.441383 0.9173468  -0.35497556 2.339389 0.7344063
#>  74:   C061  -2.5219689  3.605702 0.9092653  -1.19247902 2.289162 0.7456884
#>  75:   C086  -8.5150245  4.706661 0.8453964   1.37529591 2.372689 0.7267911
#>  76:   C071  -2.7793931  3.443594 0.9172405  -1.18719721 2.481478 0.7011633
#>  77:   C063   1.3294302  3.897822 0.8939678  -2.96340111 2.536397 0.6877894
#>  78:   C031 -13.4924994  3.730517 0.9028748   3.09961485 2.477517 0.7021166
#>  79:   C035  -5.8742194  3.749971 0.9018592  -0.23327271 2.457564 0.7068953
#>  80:   C051  -2.1297695  3.700997 0.9044059  -2.36740751 2.440174 0.7110287
#>  81:   C060 -13.5520826  6.852941 0.6722462   2.23717228 2.579745 0.6770266
#>  82:   C003 -13.2013990  5.433674 0.7939461   1.94251059 2.442075 0.7105782
#>  83:   C007  -1.2481689  3.612410 0.9089274  -3.28628137 2.611776 0.6689566
#>  84:   C021 -17.0669334  5.091737 0.8190637   3.38820032 2.276564 0.7484798
#>  85:   C036 -11.1301795  4.267782 0.8728846   0.80807172 2.401419 0.7201348
#>  86:   C041   8.9231029  3.665539 0.9062288  -7.79606092 2.589846 0.6744926
#>  87:   C081   3.6482350  3.564611 0.9113216  -5.60540972 2.498225 0.6971160
#>  88:   C002  -8.0220146  3.458024 0.9165455  -0.68092057 2.457138 0.7069968
#>  89:   C015  -0.7109968  6.511596 0.7040838  -3.97026225 2.477670 0.7020797
#>  90:   C059  -2.9673707  3.786698 0.8999274  -3.12750744 2.493090 0.6983600
#>  91:   C026 -14.6733351  4.093797 0.8830376   1.88981674 2.423228 0.7150283
#>  92:   C090 -18.5677349  4.174761 0.8783654   2.80401582 2.386401 0.7236241
#>  93:   C040 -12.5826274  4.657878 0.8485846   0.16575960 2.386793 0.7235335
#>  94:   C017  -2.9687239  3.459122 0.9164925  -3.97866568 2.427182 0.7140976
#>  95:   C052 -12.9669835  4.051819 0.8854240   0.25101011 2.498160 0.6971320
#>  96:   C100 -11.7362864  3.614629 0.9088154  -0.37513762 2.323516 0.7379981
#>  97:   C024  -8.8179680  3.764385 0.9011033  -2.15000883 2.441655 0.7106779
#>  98:   C018 -13.3968346  4.331309 0.8690721  -1.21215364 2.535050 0.6881209
#>  99:   C044  -5.6775002  4.146094 0.8800302  -4.72035710 2.544265 0.6858496
#> 100:   C053 -10.8967751  4.144844 0.8801025  -2.78307462 2.429144 0.7136352
#>       clone         DGE    se.DGE   rel.DGE          IGE   se.IGE   rel.IGE
#>            class         TGV
#>           <char>       <num>
#>   1: Homeostatic  22.7522784
#>   2: Homeostatic  18.3161186
#>   3: Homeostatic  14.6848165
#>   4:  Aggressive  12.1039669
#>   5: Homeostatic  11.6414199
#>   6:  Aggressive  11.5200943
#>   7:   Sensitive  11.3668196
#>   8: Homeostatic  11.3304979
#>   9: Homeostatic  10.9976479
#>  10: Homeostatic  10.6777972
#>  11:   Sensitive   9.8638740
#>  12:   Sensitive   9.6148714
#>  13:  Aggressive   9.4322046
#>  14:  Aggressive   8.2377365
#>  15: Homeostatic   7.7480201
#>  16:   Sensitive   7.2475296
#>  17: Homeostatic   7.1532525
#>  18: Homeostatic   6.1284828
#>  19:   Sensitive   5.7933625
#>  20:   Sensitive   5.7462839
#>  21: Homeostatic   5.5568479
#>  22: Homeostatic   5.2543344
#>  23:  Aggressive   5.2517821
#>  24: Homeostatic   4.8905237
#>  25:   Sensitive   4.8609054
#>  26:   Sensitive   4.5858883
#>  27:   Sensitive   4.4586873
#>  28: Homeostatic   4.2405798
#>  29: Homeostatic   3.9850640
#>  30: Homeostatic   3.7157477
#>  31: Homeostatic   3.6348498
#>  32: Homeostatic   3.5797859
#>  33:   Sensitive   3.5231346
#>  34:   Sensitive   3.5104428
#>  35:  Aggressive   3.4989442
#>  36:   Sensitive   2.9279535
#>  37:  Aggressive   2.6742017
#>  38: Homeostatic   2.5804972
#>  39: Homeostatic   2.4788536
#>  40:  Aggressive   2.4301318
#>  41: Homeostatic   2.2669995
#>  42: Homeostatic   2.1683803
#>  43: Homeostatic   2.0721697
#>  44:  Aggressive   1.9301770
#>  45: Homeostatic   1.8757668
#>  46: Homeostatic   1.8688058
#>  47: Homeostatic   1.4203543
#>  48:   Sensitive   1.2267796
#>  49: Homeostatic   1.0468218
#>  50: Homeostatic   0.7993871
#>  51: Homeostatic   0.4939297
#>  52: Homeostatic   0.3643810
#>  53: Homeostatic   0.2397196
#>  54:   Sensitive   0.1410357
#>  55: Homeostatic  -0.2917035
#>  56: Homeostatic  -0.3440810
#>  57:   Sensitive  -0.6571185
#>  58: Homeostatic  -0.7631467
#>  59: Homeostatic  -1.3702285
#>  60: Homeostatic  -1.5662005
#>  61: Homeostatic  -1.6405278
#>  62: Homeostatic  -1.7242337
#>  63:  Aggressive  -2.7918368
#>  64: Homeostatic  -2.9553596
#>  65: Homeostatic  -3.0021987
#>  66:   Sensitive  -3.0699578
#>  67: Homeostatic  -3.3159248
#>  68:  Aggressive  -3.9215519
#>  69: Homeostatic  -4.2828320
#>  70: Homeostatic  -4.3852658
#>  71: Homeostatic  -4.3867444
#>  72:   Sensitive  -4.5485533
#>  73: Homeostatic  -4.8311086
#>  74: Homeostatic  -5.3022113
#>  75: Homeostatic  -5.3085481
#>  76: Homeostatic  -5.5473210
#>  77: Homeostatic  -5.5796836
#>  78: Homeostatic  -6.2658059
#>  79: Homeostatic  -6.4180904
#>  80: Homeostatic  -7.6493355
#>  81: Homeostatic  -8.3361576
#>  82: Homeostatic  -8.6724722
#>  83: Homeostatic  -8.9100719
#>  84: Homeostatic  -9.1674083
#>  85: Homeostatic  -9.2461755
#>  86:  Aggressive  -9.2532661
#>  87:  Aggressive  -9.4206720
#>  88: Homeostatic  -9.6095681
#>  89:  Aggressive  -9.9675884
#>  90: Homeostatic -10.2590953
#>  91: Homeostatic -10.2672630
#>  92: Homeostatic -12.0302249
#>  93: Homeostatic -12.1961620
#>  94:  Aggressive -12.2449078
#>  95: Homeostatic -12.3817582
#>  96: Homeostatic -12.6109126
#>  97: Homeostatic -13.8306731
#>  98: Homeostatic -16.2229480
#>  99:  Aggressive -16.6829238
#> 100: Homeostatic -17.3854611
#>            class         TGV
 print(results, category = 'summar')
#> ===> Likelihood ratio tests
#>    effect  LR-statistic    Pr(Chisq)
#> 1     DGE  1.098212e+02 0.000000e+00
#> 2     IGE  5.671546e+01 2.520206e-14
#> 3 DGE:age  2.022865e+00 7.747327e-02
#> 4 IGE:age -9.514352e-05 5.000000e-01
#> ===> Variance components
#>                                   component   std.error    z.ratio bound %ch
#> block:age                      1.621697e+01  5.83666336  2.7784659     P 0.0
#> cor(IGE_DGE)                  -5.829904e-01  0.10731046 -5.4327456     U 0.0
#> DGE                            1.432868e+02 23.51414742  6.0936427     P 0.0
#> IGE                            2.060568e+01  4.54000889  4.5386865     P 0.0
#> DGE:age                        5.282378e+00  4.47060368  1.1815805     P 0.2
#> IGE:age                        1.477966e-06          NA         NA     B  NA
#> R=area_A1:age_3y               1.776881e+02 14.08064019 12.6193183     P 0.0
#> R=autocor(row):area_A1:age_3y  7.838337e-02  0.06253917  1.2533485     U 0.1
#> R=autocor(col):area_A1:age_3y -1.413340e-01  0.06366699 -2.2198942     U 0.0
#> R=area_A3:age_3y               1.335698e+02 10.23241149 13.0536001     P 0.0
#> R=autocor(row):area_A3:age_3y  3.150012e-02  0.06750273  0.4666495     U 0.8
#> R=autocor(col):area_A3:age_3y  7.145465e-02  0.06365827  1.1224724     U 0.1
#> R=area_A1:age_6y               4.260002e+02 34.06428764 12.5057708     P 0.0
#> R=autocor(row):area_A1:age_6y -3.134735e-02  0.06819205 -0.4596921     U 0.6
#> R=autocor(col):area_A1:age_6y -1.942322e-01  0.06383635 -3.0426582     U 0.0
#> R=area_A3:age_6y               2.441681e+02 19.30333795 12.6490090     P 0.0
#> R=autocor(row):area_A3:age_6y -3.653100e-02  0.07507359 -0.4866026     U 0.7
#> R=autocor(col):area_A3:age_6y  7.585456e-02  0.06931897  1.0942829     U 0.2
 print(results, category = 'blup.main')
#> ===> DGE, IGE and TGV
#>       clone         DGE    se.DGE   rel.DGE          IGE   se.IGE   rel.IGE
#>      <fctr>       <num>     <num>     <num>        <num>    <num>     <num>
#>   1:   C099  24.6155934  3.577760 0.9106661  -0.79919798 2.460610 0.7061682
#>   2:   C087  13.2013688  4.119508 0.8815638   2.19377705 2.477019 0.7022363
#>   3:   C083  22.7097354  3.620486 0.9085197  -3.44198318 2.510407 0.6941551
#>   4:   C028  26.1498864  3.475277 0.9157107  -6.02446200 2.636438 0.6626753
#>   5:   C066   8.5283065  3.635331 0.9077680   1.33525136 2.334376 0.7355431
#>   6:   C092  28.1174949  3.666496 0.9061798  -7.11882264 2.645961 0.6602338
#>   7:   C025   0.3828506  3.581798 0.9104644   4.71115499 2.388269 0.7231915
#>   8:   C073  14.4757695  3.552133 0.9119413  -1.34904442 2.389080 0.7230034
#>   9:   C023  11.1799264  4.429750 0.8630531  -0.07818138 2.651615 0.6587803
#>  10:   C096   8.4122922  3.420055 0.9183681   0.97170205 2.498207 0.6971205
#>  11:   C033   0.9581647  3.997578 0.8884710   3.81976471 2.434505 0.7123699
#>  12:   C019  -1.7493381  5.512207 0.7879468   4.87424467 2.372559 0.7268210
#>  13:   C075  34.7749389  3.446719 0.9170903 -10.86980030 2.542849 0.6861990
#>  14:   C093  26.9481426  3.548553 0.9121187  -8.02511580 2.543072 0.6861440
#>  15:   C085   7.1142751  3.662342 0.9063923   0.27182076 2.525885 0.6903720
#>  16:   C076 -14.0265744 10.313925 0.2575935   9.12471633 2.237582 0.7570198
#>  17:   C046   3.6440155  3.469133 0.9160084   1.50515350 2.546888 0.6852015
#>  18:   C011   8.8306228  3.819568 0.8981825  -1.15897999 2.523569 0.6909396
#>  19:   C039 -12.0444837  4.058322 0.8850559   7.65086450 2.476385 0.7023886
#>  20:   C082  -3.6719377  3.551550 0.9119702   4.03958734 2.468875 0.7041910
#>  21:   C080   1.3295020  3.453747 0.9167518   1.81315897 2.522621 0.6911717
#>  22:   C008   0.6641248  3.423279 0.9182141   1.96879549 2.453173 0.7079417
#>  23:   C097  16.1257274  3.620008 0.9085439  -4.66396455 2.530240 0.6893034
#>  24:   C089  10.9336620  3.433417 0.9177290  -2.59197392 2.515964 0.6927996
#>  25:   C004  -8.1110852  4.080600 0.8837905   5.56384109 2.424791 0.7146605
#>  26:   C037  -8.8753419 10.513700 0.2285551   5.77368180 2.601459 0.6715667
#>  27:   C016 -10.5108365  6.807094 0.6766169   6.42060688 2.435875 0.7120460
#>  28:   C054   4.6542420  3.725631 0.9031291  -0.17742466 2.437575 0.7116439
#>  29:   C022  -3.3588666  3.811756 0.8985986   3.14989920 2.659521 0.6567426
#>  30:   C020   1.0856965  4.067402 0.8845409   1.12806025 2.426654 0.7142221
#>  31:   C038  12.7536946  3.685042 0.9052283  -3.91118102 2.600725 0.6717520
#>  32:   C029   9.1263106  3.929047 0.8922622  -2.37897041 2.632571 0.6636640
#>  33:   C048  -7.2699648  4.140305 0.8803650   4.62928876 2.518965 0.6920662
#>  34:   C030  -9.8780638  4.650783 0.8490455   5.74248980 2.312698 0.7404322
#>  35:   C091  16.3893815  3.722072 0.9033141  -5.52886191 2.467639 0.7044872
#>  36:   C042 -10.3757947  4.306107 0.8705913   5.70613591 2.532117 0.6888422
#>  37:   C094  12.7981424  3.556092 0.9117449  -4.34227862 2.568570 0.6798186
#>  38:   C012   1.0520764  3.961655 0.8904665   0.65555788 2.662183 0.6560551
#>  39:   C006   5.7488966  3.660083 0.9065077  -1.40256034 2.497139 0.6973793
#>  40:   C027  17.7637866  4.127863 0.8810829  -6.57678701 2.661813 0.6561507
#>  41:   C058  -4.4052486  4.538619 0.8562389   2.86180659 2.573175 0.6786698
#>  42:   C013  -4.4145988  3.830370 0.8976058   2.82351806 2.548220 0.6848720
#>  43:   C014   0.6429174  3.439070 0.9174578   0.61302333 2.495953 0.6976667
#>  44:   C055  20.3115351  3.588649 0.9101215  -7.88398321 2.639078 0.6619993
#>  45:   C098   8.9223245  3.825352 0.8978739  -3.02235243 2.395537 0.7215039
#>  46:   C010   4.4739674  3.521331 0.9134619  -1.11738481 2.609469 0.6695412
#>  47:   C057   9.2017902  3.533404 0.9128674  -3.33755044 2.549294 0.6846063
#>  48:   C069  -8.8977865  4.105342 0.8823769   4.34254687 2.798305 0.6199827
#>  49:   C078   0.2608237  3.607569 0.9091713   0.33712395 2.587089 0.6751851
#>  50:   C049   1.1412872  3.986351 0.8890966  -0.14664503 2.523758 0.6908933
#>  51:   C047  -0.9559314 10.559552 0.2218116   0.62186268 2.678899 0.6517222
#>  52:   C045  -7.6352539  4.138562 0.8804656   3.43113860 2.492540 0.6984930
#>  53:   C062   1.6158885  3.537829 0.9126491  -0.59025523 2.393447 0.7219899
#>  54:   C095 -13.8146351  4.668129 0.8479174   5.98575328 2.480467 0.7014068
#>  55:   C001   4.5943942  3.481016 0.9154321  -2.09570546 2.585117 0.6756802
#>  56:   C050   0.6659203 10.396883 0.2456028  -0.43320157 2.394549 0.7217338
#>  57:   C079 -11.7117120  4.362077 0.8672054   4.74144669 2.423588 0.7149436
#>  58:   C032  -3.4555615  3.835731 0.8973190   1.15480877 2.588327 0.6748742
#>  59:   C056   2.2619037  4.020630 0.8871811  -1.55786473 2.413082 0.7174098
#>  60:   C084   4.8118781  3.659456 0.9065398  -2.73563382 2.465959 0.7048895
#>  61:   C067   3.8597666  3.407732 0.9189553  -2.35914171 2.472562 0.7033068
#>  62:   C070   1.1098787  3.929629 0.8922303  -1.21558453 2.475821 0.7025242
#>  63:   C088   8.8995446  3.359176 0.9212484  -5.01457255 2.403720 0.7195980
#>  64:   C065  -9.7579765  3.937199 0.8918146   2.91772336 2.591672 0.6740334
#>  65:   C034  -6.1914023  4.461427 0.8610875   1.36788737 2.450179 0.7086542
#>  66:   C043 -12.9939921  6.487007 0.7063145   4.25653639 2.292927 0.7448513
#>  67:   C074   0.9348632  3.524303 0.9133157  -1.82321355 2.369375 0.7275539
#>  68:   C009   9.0797815  3.782424 0.9001532  -5.57642663 2.658715 0.6569505
#>  69:   C064  -4.3933717  3.552215 0.9119372   0.04741177 2.603557 0.6710367
#>  70:   C077 -10.4876080  3.957612 0.8906899   2.61736719 2.575751 0.6780259
#>  71:   C072  -0.3063273  3.349451 0.9217037  -1.75013945 2.423955 0.7148573
#>  72:   C005 -13.7764233  3.817664 0.8982840   3.95794324 2.507961 0.6947506
#>  73:   C068  -4.0034897  3.441383 0.9173468  -0.35497556 2.339389 0.7344063
#>  74:   C061  -2.5219689  3.605702 0.9092653  -1.19247902 2.289162 0.7456884
#>  75:   C086  -8.5150245  4.706661 0.8453964   1.37529591 2.372689 0.7267911
#>  76:   C071  -2.7793931  3.443594 0.9172405  -1.18719721 2.481478 0.7011633
#>  77:   C063   1.3294302  3.897822 0.8939678  -2.96340111 2.536397 0.6877894
#>  78:   C031 -13.4924994  3.730517 0.9028748   3.09961485 2.477517 0.7021166
#>  79:   C035  -5.8742194  3.749971 0.9018592  -0.23327271 2.457564 0.7068953
#>  80:   C051  -2.1297695  3.700997 0.9044059  -2.36740751 2.440174 0.7110287
#>  81:   C060 -13.5520826  6.852941 0.6722462   2.23717228 2.579745 0.6770266
#>  82:   C003 -13.2013990  5.433674 0.7939461   1.94251059 2.442075 0.7105782
#>  83:   C007  -1.2481689  3.612410 0.9089274  -3.28628137 2.611776 0.6689566
#>  84:   C021 -17.0669334  5.091737 0.8190637   3.38820032 2.276564 0.7484798
#>  85:   C036 -11.1301795  4.267782 0.8728846   0.80807172 2.401419 0.7201348
#>  86:   C041   8.9231029  3.665539 0.9062288  -7.79606092 2.589846 0.6744926
#>  87:   C081   3.6482350  3.564611 0.9113216  -5.60540972 2.498225 0.6971160
#>  88:   C002  -8.0220146  3.458024 0.9165455  -0.68092057 2.457138 0.7069968
#>  89:   C015  -0.7109968  6.511596 0.7040838  -3.97026225 2.477670 0.7020797
#>  90:   C059  -2.9673707  3.786698 0.8999274  -3.12750744 2.493090 0.6983600
#>  91:   C026 -14.6733351  4.093797 0.8830376   1.88981674 2.423228 0.7150283
#>  92:   C090 -18.5677349  4.174761 0.8783654   2.80401582 2.386401 0.7236241
#>  93:   C040 -12.5826274  4.657878 0.8485846   0.16575960 2.386793 0.7235335
#>  94:   C017  -2.9687239  3.459122 0.9164925  -3.97866568 2.427182 0.7140976
#>  95:   C052 -12.9669835  4.051819 0.8854240   0.25101011 2.498160 0.6971320
#>  96:   C100 -11.7362864  3.614629 0.9088154  -0.37513762 2.323516 0.7379981
#>  97:   C024  -8.8179680  3.764385 0.9011033  -2.15000883 2.441655 0.7106779
#>  98:   C018 -13.3968346  4.331309 0.8690721  -1.21215364 2.535050 0.6881209
#>  99:   C044  -5.6775002  4.146094 0.8800302  -4.72035710 2.544265 0.6858496
#> 100:   C053 -10.8967751  4.144844 0.8801025  -2.78307462 2.429144 0.7136352
#>       clone         DGE    se.DGE   rel.DGE          IGE   se.IGE   rel.IGE
#>            class         TGV
#>           <char>       <num>
#>   1: Homeostatic  22.7522784
#>   2: Homeostatic  18.3161186
#>   3: Homeostatic  14.6848165
#>   4:  Aggressive  12.1039669
#>   5: Homeostatic  11.6414199
#>   6:  Aggressive  11.5200943
#>   7:   Sensitive  11.3668196
#>   8: Homeostatic  11.3304979
#>   9: Homeostatic  10.9976479
#>  10: Homeostatic  10.6777972
#>  11:   Sensitive   9.8638740
#>  12:   Sensitive   9.6148714
#>  13:  Aggressive   9.4322046
#>  14:  Aggressive   8.2377365
#>  15: Homeostatic   7.7480201
#>  16:   Sensitive   7.2475296
#>  17: Homeostatic   7.1532525
#>  18: Homeostatic   6.1284828
#>  19:   Sensitive   5.7933625
#>  20:   Sensitive   5.7462839
#>  21: Homeostatic   5.5568479
#>  22: Homeostatic   5.2543344
#>  23:  Aggressive   5.2517821
#>  24: Homeostatic   4.8905237
#>  25:   Sensitive   4.8609054
#>  26:   Sensitive   4.5858883
#>  27:   Sensitive   4.4586873
#>  28: Homeostatic   4.2405798
#>  29: Homeostatic   3.9850640
#>  30: Homeostatic   3.7157477
#>  31: Homeostatic   3.6348498
#>  32: Homeostatic   3.5797859
#>  33:   Sensitive   3.5231346
#>  34:   Sensitive   3.5104428
#>  35:  Aggressive   3.4989442
#>  36:   Sensitive   2.9279535
#>  37:  Aggressive   2.6742017
#>  38: Homeostatic   2.5804972
#>  39: Homeostatic   2.4788536
#>  40:  Aggressive   2.4301318
#>  41: Homeostatic   2.2669995
#>  42: Homeostatic   2.1683803
#>  43: Homeostatic   2.0721697
#>  44:  Aggressive   1.9301770
#>  45: Homeostatic   1.8757668
#>  46: Homeostatic   1.8688058
#>  47: Homeostatic   1.4203543
#>  48:   Sensitive   1.2267796
#>  49: Homeostatic   1.0468218
#>  50: Homeostatic   0.7993871
#>  51: Homeostatic   0.4939297
#>  52: Homeostatic   0.3643810
#>  53: Homeostatic   0.2397196
#>  54:   Sensitive   0.1410357
#>  55: Homeostatic  -0.2917035
#>  56: Homeostatic  -0.3440810
#>  57:   Sensitive  -0.6571185
#>  58: Homeostatic  -0.7631467
#>  59: Homeostatic  -1.3702285
#>  60: Homeostatic  -1.5662005
#>  61: Homeostatic  -1.6405278
#>  62: Homeostatic  -1.7242337
#>  63:  Aggressive  -2.7918368
#>  64: Homeostatic  -2.9553596
#>  65: Homeostatic  -3.0021987
#>  66:   Sensitive  -3.0699578
#>  67: Homeostatic  -3.3159248
#>  68:  Aggressive  -3.9215519
#>  69: Homeostatic  -4.2828320
#>  70: Homeostatic  -4.3852658
#>  71: Homeostatic  -4.3867444
#>  72:   Sensitive  -4.5485533
#>  73: Homeostatic  -4.8311086
#>  74: Homeostatic  -5.3022113
#>  75: Homeostatic  -5.3085481
#>  76: Homeostatic  -5.5473210
#>  77: Homeostatic  -5.5796836
#>  78: Homeostatic  -6.2658059
#>  79: Homeostatic  -6.4180904
#>  80: Homeostatic  -7.6493355
#>  81: Homeostatic  -8.3361576
#>  82: Homeostatic  -8.6724722
#>  83: Homeostatic  -8.9100719
#>  84: Homeostatic  -9.1674083
#>  85: Homeostatic  -9.2461755
#>  86:  Aggressive  -9.2532661
#>  87:  Aggressive  -9.4206720
#>  88: Homeostatic  -9.6095681
#>  89:  Aggressive  -9.9675884
#>  90: Homeostatic -10.2590953
#>  91: Homeostatic -10.2672630
#>  92: Homeostatic -12.0302249
#>  93: Homeostatic -12.1961620
#>  94:  Aggressive -12.2449078
#>  95: Homeostatic -12.3817582
#>  96: Homeostatic -12.6109126
#>  97: Homeostatic -13.8306731
#>  98: Homeostatic -16.2229480
#>  99:  Aggressive -16.6829238
#> 100: Homeostatic -17.3854611
#>            class         TGV
 print(results, category = 'blup.within')
#> ===> DGE, IGE and TGV (all ages)
#>       clone    age        DGE   se.DGE   rel.DGE        IGE      se.IGE
#>      <char> <char>      <num>    <num>     <num>      <num>       <num>
#>   1:   C099     6y  27.029438 2.172472 0.9670616 -0.7991979 0.001215716
#>   2:   C099     3y  23.506051 2.158921 0.9674713 -0.7991975 0.001215716
#>   3:   C087     6y  14.493990 2.207611 0.9659875  2.1937774 0.001215716
#>   4:   C087     3y  12.833064 2.193758 0.9664130  2.1937772 0.001215716
#>   5:   C083     6y  23.676494 2.175711 0.9669633 -3.4419833 0.001215716
#>  ---                                                                   
#> 196:   C044     3y  -5.192635 2.201943 0.9661619 -4.7203577 0.001215716
#> 197:   C018     3y -14.062931 2.210587 0.9658957 -1.2121537 0.001215716
#> 198:   C053     6y -11.257675 2.212328 0.9658420 -2.7830745 0.001215716
#> 199:   C044     6y  -6.883679 2.221672 0.9655528 -4.7203571 0.001215716
#> 200:   C053     3y -11.382185 2.198336 0.9662727 -2.7830753 0.001215716
#>        rel.IGE       class       TGV
#>          <num>      <char>     <num>
#>   1: 0.9999999 Homeostatic  25.19806
#>   2: 0.9999999  Aggressive  21.61080
#>   3: 0.9999999 Homeostatic  19.52108
#>   4: 0.9999999 Homeostatic  18.03547
#>   5: 0.9999999 Homeostatic  15.78911
#>  ---                                
#> 196: 0.9999999 Homeostatic -16.38668
#> 197: 0.9999999 Homeostatic -16.93748
#> 198: 0.9999999 Homeostatic -17.63516
#> 199: 0.9999999  Aggressive -17.70049
#> 200: 0.9999999   Sensitive -17.98208
 print(results, category = 'blup.within', age = '6y')
#> ===> DGE, IGE and TGV(6y)
#>       clone    age         DGE   se.DGE   rel.DGE          IGE      se.IGE
#>      <char> <char>       <num>    <num>     <num>        <num>       <num>
#>   1:   C099     6y  27.0294376 2.172472 0.9670616  -0.79919794 0.001215716
#>   2:   C087     6y  14.4939900 2.207611 0.9659875   2.19377745 0.001215716
#>   3:   C083     6y  23.6764945 2.175711 0.9669633  -3.44198329 0.001215716
#>   4:   C028     6y  27.1546610 2.159342 0.9674586  -6.02446202 0.001215716
#>   5:   C075     6y  38.2217292 2.159263 0.9674610 -10.86980055 0.001215716
#>   6:   C092     6y  29.4694007 2.179840 0.9668378  -7.11882278 0.001215716
#>   7:   C073     6y  15.4748916 2.174300 0.9670062  -1.34904430 0.001215716
#>   8:   C066     6y   9.3236155 2.180907 0.9668053   1.33525160 0.001215716
#>   9:   C023     6y  11.9733495 2.224874 0.9654534  -0.07818142 0.001215716
#>  10:   C096     6y   8.9665315 2.162160 0.9673736   0.97170210 0.001215716
#>  11:   C025     6y   0.3877892 2.182578 0.9667545   4.71115523 0.001215716
#>  12:   C033     6y   1.7190281 2.208970 0.9659456   3.81976486 0.001215716
#>  13:   C093     6y  28.3993057 2.172588 0.9670581  -8.02511615 0.001215716
#>  14:   C019     6y  -1.5481778 2.256954 0.9644500   4.87424515 0.001215716
#>  15:   C085     6y   7.4807318 2.179911 0.9668357   0.27182101 0.001215716
#>  16:   C046     6y   3.9565075 2.162320 0.9673687   1.50515332 0.001215716
#>  17:   C011     6y   9.8009524 2.192499 0.9664515  -1.15897989 0.001215716
#>  18:   C076     6y -14.0265744 2.296269 0.9632007   9.12471680 0.001215716
#>  19:   C097     6y  17.0096315 2.176737 0.9669322  -4.66396469 0.001215716
#>  20:   C080     6y   1.2079222 2.158692 0.9674782   1.81315916 0.001215716
#>  21:   C008     6y   0.8271023 2.158616 0.9674804   1.96879560 0.001215716
#>  22:   C082     6y  -3.9959414 2.174063 0.9670134   4.03958744 0.001215716
#>  23:   C038     6y  13.8646061 2.180443 0.9668195  -3.91118123 0.001215716
#>  24:   C039     6y -12.6460691 2.216245 0.9657209   7.65086493 0.001215716
#>  25:   C004     6y  -7.9607236 2.223533 0.9654951   5.56384160 0.001215716
#>  26:   C091     6y  17.3968534 2.189911 0.9665307  -5.52886196 0.001215716
#>  27:   C016     6y -10.3083173 2.283544 0.9636075   6.42060697 0.001215716
#>  28:   C054     6y   4.8051267 2.186872 0.9666235  -0.17742475 0.001215716
#>  29:   C037     6y  -8.8753419 2.296269 0.9632007   5.77368218 0.001215716
#>  30:   C022     6y  -3.2746308 2.196783 0.9663203   3.14989915 0.001215716
#>  31:   C089     6y   9.6697871 2.157963 0.9675001  -2.59197406 0.001215716
#>  32:   C094     6y  13.4964277 2.172179 0.9670705  -4.34227906 0.001215716
#>  33:   C029     6y   8.8714633 2.199444 0.9662387  -2.37897033 0.001215716
#>  34:   C030     6y  -9.8259844 2.250739 0.9646456   5.74249008 0.001215716
#>  35:   C027     6y  18.3520002 2.212124 0.9658483  -6.57678723 0.001215716
#>  36:   C020     6y   0.4872292 2.216155 0.9657237   1.12806038 0.001215716
#>  37:   C048     6y  -7.6881195 2.217823 0.9656721   4.62928889 0.001215716
#>  38:   C042     6y -10.3380963 2.225769 0.9654257   5.70613646 0.001215716
#>  39:   C098     6y   9.5264183 2.204017 0.9660981  -3.02235238 0.001215716
#>  40:   C014     6y   1.0114863 2.159360 0.9674580   0.61302368 0.001215716
#>  41:   C006     6y   5.5998429 2.180201 0.9668268  -1.40256034 0.001215716
#>  42:   C010     6y   4.7611608 2.166062 0.9672557  -1.11738501 0.001215716
#>  43:   C012     6y   0.3695906 2.199983 0.9662221   0.65555829 0.001215716
#>  44:   C057     6y   9.4629144 2.165518 0.9672722  -3.33755055 0.001215716
#>  45:   C058     6y  -5.0105925 2.228600 0.9653377   2.86180686 0.001215716
#>  46:   C055     6y  19.4600262 2.173077 0.9670433  -7.88398350 0.001215716
#>  47:   C001     6y   5.6619727 2.162375 0.9673671  -2.09570559 0.001215716
#>  48:   C013     6y  -5.6338815 2.191711 0.9664756   2.82351801 0.001215716
#>  49:   C069     6y  -9.2067488 2.214239 0.9657829   4.34254699 0.001215716
#>  50:   C047     6y  -0.9559314 2.296269 0.9632007   0.62186289 0.001215716
#>  51:   C045     6y  -7.8082514 2.219747 0.9656125   3.43113904 0.001215716
#>  52:   C049     6y   0.3638449 2.199740 0.9662296  -0.14664502 0.001215716
#>  53:   C062     6y   1.3287742 2.168796 0.9671730  -0.59025528 0.001215716
#>  54:   C078     6y  -1.0920746 2.175823 0.9669599   0.33712399 0.001215716
#>  55:   C050     6y   0.6659203 2.296269 0.9632007  -0.43320159 0.001215716
#>  56:   C095     6y -14.5934472 2.233451 0.9651866   5.98575361 0.001215716
#>  57:   C070     6y   1.8830341 2.212265 0.9658439  -1.21558455 0.001215716
#>  58:   C032     6y  -3.6088782 2.197059 0.9663118   1.15480898 0.001215716
#>  59:   C056     6y   2.3345703 2.202555 0.9661431  -1.55786460 0.001215716
#>  60:   C079     6y -12.1551033 2.226101 0.9654153   4.74144695 0.001215716
#>  61:   C084     6y   4.3274785 2.180151 0.9668284  -2.73563377 0.001215716
#>  62:   C067     6y   2.6303669 2.151483 0.9676950  -2.35914180 0.001215716
#>  63:   C074     6y   1.1926201 2.168644 0.9671776  -1.82321388 0.001215716
#>  64:   C009     6y   9.6988826 2.187966 0.9665901  -5.57642668 0.001215716
#>  65:   C088     6y   8.3666320 2.152061 0.9676776  -5.01457257 0.001215716
#>  66:   C043     6y -13.0764058 2.272870 0.9639469   4.25653670 0.001215716
#>  67:   C065     6y -10.0373586 2.201023 0.9661902   2.91772338 0.001215716
#>  68:   C034     6y  -6.5928179 2.223541 0.9654948   1.36788733 0.001215716
#>  69:   C064     6y  -4.2415910 2.165925 0.9672598   0.04741174 0.001215716
#>  70:   C077     6y -10.8128028 2.199713 0.9662304   2.61736714 0.001215716
#>  71:   C068     6y  -4.1397853 2.162676 0.9673580  -0.35497554 0.001215716
#>  72:   C071     6y  -2.5724689 2.162032 0.9673775  -1.18719735 0.001215716
#>  73:   C072     6y  -1.3548142 2.151291 0.9677008  -1.75013964 0.001215716
#>  74:   C086     6y  -8.5422817 2.240821 0.9649564   1.37529607 0.001215716
#>  75:   C061     6y  -2.9157133 2.181545 0.9667859  -1.19247901 0.001215716
#>  76:   C005     6y -14.7371031 2.197582 0.9662958   3.95794338 0.001215716
#>  77:   C063     6y  -0.1930758 2.235255 0.9651303  -2.96340114 0.001215716
#>  78:   C031     6y -14.0966881 2.183166 0.9667366   3.09961469 0.001215716
#>  79:   C035     6y  -6.4733368 2.187988 0.9665894  -0.23327281 0.001215716
#>  80:   C051     6y  -2.2676304 2.185000 0.9666807  -2.36740758 0.001215716
#>  81:   C060     6y -13.8539592 2.283159 0.9636197   2.23717225 0.001215716
#>  82:   C007     6y  -1.3378481 2.172737 0.9670536  -3.28628165 0.001215716
#>  83:   C041     6y   8.8436517 2.181826 0.9667774  -7.79606122 0.001215716
#>  84:   C081     6y   3.4449007 2.172944 0.9670473  -5.60540988 0.001215716
#>  85:   C003     6y -13.8529524 2.262198 0.9642846   1.94251049 0.001215716
#>  86:   C021     6y -17.1718467 2.282894 0.9636281   3.38820042 0.001215716
#>  87:   C036     6y -11.4103147 2.262723 0.9642681   0.80807171 0.001215716
#>  88:   C015     6y  -0.7843098 2.272619 0.9639548  -3.97026254 0.001215716
#>  89:   C002     6y  -8.7113064 2.162486 0.9673637  -0.68092051 0.001215716
#>  90:   C059     6y  -3.3165128 2.191668 0.9664770  -3.12750759 0.001215716
#>  91:   C026     6y -15.1317280 2.207213 0.9659997   1.88981667 0.001215716
#>  92:   C040     6y -12.4708920 2.256818 0.9644543   0.16575942 0.001215716
#>  93:   C090     6y -18.6810913 2.220425 0.9655915   2.80401604 0.001215716
#>  94:   C017     6y  -3.6065374 2.168999 0.9671668  -3.97866605 0.001215716
#>  95:   C052     6y -13.6342053 2.230359 0.9652829   0.25101014 0.001215716
#>  96:   C100     6y -12.6723577 2.175561 0.9669679  -0.37513783 0.001215716
#>  97:   C024     6y  -8.8974140 2.189143 0.9665542  -2.15000906 0.001215716
#>  98:   C018     6y -13.5818008 2.232550 0.9652147  -1.21215403 0.001215716
#>  99:   C053     6y -11.2576749 2.212328 0.9658420  -2.78307451 0.001215716
#> 100:   C044     6y  -6.8836789 2.221672 0.9655528  -4.72035713 0.001215716
#>       clone    age         DGE   se.DGE   rel.DGE          IGE      se.IGE
#>        rel.IGE       class          TGV
#>          <num>      <char>        <num>
#>   1: 0.9999999 Homeostatic  25.19805687
#>   2: 0.9999999 Homeostatic  19.52108215
#>   3: 0.9999999 Homeostatic  15.78910955
#>   4: 0.9999999 Homeostatic  13.34946585
#>   5: 0.9999999 Homeostatic  13.31332800
#>   6: 0.9999999 Homeostatic  13.15645244
#>   7: 0.9999999 Homeostatic  12.38352521
#>   8: 0.9999999 Homeostatic  12.38337565
#>   9: 0.9999999 Homeostatic  11.79419491
#>  10: 0.9999999  Aggressive  11.19320948
#>  11: 0.9999999 Homeostatic  11.18351114
#>  12: 0.9999999 Homeostatic  10.47210829
#>  13: 0.9999999 Homeostatic  10.00956509
#>  14: 0.9999999 Homeostatic   9.62126853
#>  15: 0.9999999 Homeostatic   8.10361596
#>  16: 0.9999999  Aggressive   7.40560145
#>  17: 0.9999999 Homeostatic   7.14512294
#>  18: 0.9999999 Homeostatic   6.88292681
#>  19: 0.9999999  Aggressive   6.32204775
#>  20: 0.9999999 Homeostatic   5.36281861
#>  21: 0.9999999   Sensitive   5.33864326
#>  22: 0.9999999 Homeostatic   5.26086738
#>  23: 0.9999999   Sensitive   4.90204316
#>  24: 0.9999999 Homeostatic   4.88606614
#>  25: 0.9999999 Homeostatic   4.78894906
#>  26: 0.9999999   Sensitive   4.72733743
#>  27: 0.9999999 Homeostatic   4.40465316
#>  28: 0.9999999 Homeostatic   4.39855372
#>  29: 0.9999999 Homeostatic   4.35518537
#>  30: 0.9999999  Aggressive   3.94343646
#>  31: 0.9999999 Homeostatic   3.73021812
#>  32: 0.9999999  Aggressive   3.54599401
#>  33: 0.9999999 Homeostatic   3.41999732
#>  34: 0.9999999 Homeostatic   3.33306542
#>  35: 0.9999999 Homeostatic   3.28113905
#>  36: 0.9999999 Homeostatic   3.07220583
#>  37: 0.9999999 Homeostatic   2.92000391
#>  38: 0.9999999 Homeostatic   2.73764839
#>  39: 0.9999999 Homeostatic   2.60062738
#>  40: 0.9999999  Aggressive   2.41624431
#>  41: 0.9999999 Homeostatic   2.38584318
#>  42: 0.9999999 Homeostatic   2.20064700
#>  43: 0.9999999 Homeostatic   1.87181772
#>  44: 0.9999999 Homeostatic   1.81483950
#>  45: 0.9999999   Sensitive   1.54730461
#>  46: 0.9999999 Homeostatic   1.39369426
#>  47: 0.9999999 Homeostatic   0.85961452
#>  48: 0.9999999 Homeostatic   0.83627578
#>  49: 0.9999999 Homeostatic   0.74429881
#>  50: 0.9999999  Aggressive   0.46908187
#>  51: 0.9999999 Homeostatic   0.05428365
#>  52: 0.9999999 Homeostatic   0.02780446
#>  53: 0.9999999 Homeostatic  -0.02380954
#>  54: 0.9999999 Homeostatic  -0.31954708
#>  55: 0.9999999 Homeostatic  -0.32677122
#>  56: 0.9999999 Homeostatic  -0.87695332
#>  57: 0.9999999 Homeostatic  -0.90250628
#>  58: 0.9999999 Homeostatic  -0.96260652
#>  59: 0.9999999 Homeostatic  -1.23531272
#>  60: 0.9999999 Homeostatic  -1.28996717
#>  61: 0.9999999 Homeostatic  -1.94129003
#>  62: 0.9999999 Homeostatic  -2.77566149
#>  63: 0.9999999   Sensitive  -2.98531696
#>  64: 0.9999999 Homeostatic  -3.07962907
#>  65: 0.9999999   Sensitive  -3.12437792
#>  66: 0.9999999 Homeostatic  -3.32245273
#>  67: 0.9999999   Sensitive  -3.35132750
#>  68: 0.9999999 Homeostatic  -3.45827217
#>  69: 0.9999999  Aggressive  -4.13294587
#>  70: 0.9999999 Homeostatic  -4.81504504
#>  71: 0.9999999 Homeostatic  -4.95322005
#>  72: 0.9999999   Sensitive  -5.29295930
#>  73: 0.9999999  Aggressive  -5.36529997
#>  74: 0.9999999 Homeostatic  -5.39075870
#>  75: 0.9999999 Homeostatic  -5.64830676
#>  76: 0.9999999 Homeostatic  -5.66738360
#>  77: 0.9999999 Homeostatic  -6.98377853
#>  78: 0.9999999 Homeostatic  -6.99384875
#>  79: 0.9999999 Homeostatic  -7.00788691
#>  80: 0.9999999 Homeostatic  -7.69260001
#>  81: 0.9999999 Homeostatic  -8.72742682
#>  82: 0.9999999 Homeostatic  -8.86843909
#>  83: 0.9999999   Sensitive  -9.02120429
#>  84: 0.9999999 Homeostatic  -9.40002664
#>  85: 0.9999999 Homeostatic  -9.40164433
#>  86: 0.9999999   Sensitive  -9.40770645
#>  87: 0.9999999 Homeostatic  -9.55859951
#>  88: 0.9999999   Sensitive  -9.88225889
#>  89: 0.9999999   Sensitive -10.27165160
#>  90: 0.9999999 Homeostatic -10.48326936
#>  91: 0.9999999 Homeostatic -10.80116908
#>  92: 0.9999999 Homeostatic -12.09105037
#>  93: 0.9999999 Homeostatic -12.25562326
#>  94: 0.9999999 Homeostatic -12.72374332
#>  95: 0.9999999   Sensitive -13.05900971
#>  96: 0.9999999 Homeostatic -13.53199475
#>  97: 0.9999999   Sensitive -13.82420989
#>  98: 0.9999999 Homeostatic -16.35947999
#>  99: 0.9999999 Homeostatic -17.63515503
#> 100: 0.9999999  Aggressive -17.70048727
#>        rel.IGE       class          TGV
 # Note that the ages are labelled as "3y" and "6y" in the example dataset 
 # }