Extract outputs from the genetic-spatial competition model

This function provides responses extracted from the genetic-spatial competition model fitted using asr() or asr_ma().

resp(prep.out, model, weight.tgv = FALSE, sd.class = 1)

Arguments

prep.out

A comprepfor or comprepcrop object.

model

An compmod object, obtained from asr() or asr_ma() functions.

weight.tgv

A logical value or a numeric vector of size two. If TRUE, the function will use the direct and indirect genetic effects' reliability as a weight when estimating the total genotypic value. If a vector is provided, the values within the vector are used as weights. Defaults to FALSE, so no weight is used. See Details for more information on these methods.

sd.class

A numeric value defining the weight to establish competition classes. Defaults to 1. See Details for more information.

Value

The function returns:

lrt : a data frame with the Likelihood Ratio Test results. Available if lrtest = TRUE in asr().

varcomp : A data frame summarizing the random parameter vector (object$vparameters). Variance component ratios are included if param = "gamma", and a measure of precision (standard error) is included along with boundary constraints at termination and the percentage change in the final iteration.

heritability : a matrix (or a data frame, if a model with heterogeneous residual variances is fitted) containing the direct genetic effects heritability and the total heritability (see details). Available if cor = TRUE in asr() or asr_ma().

blups : A list with a data frame containing the direct (DGE) and indirect genetic effects (IGE), their standard errors, the competition class of each genotype and the total genotypic value (TGV). If a multi-age model was fitted, blup will have a further data frame with the within-ages DGE and IGE. If other random effects were declared in the model, blup will contain a last data frame with the corresponding BLUPs.

Details

The genetic competition model provides the direct (DGE) and indirect genetic effects (IGE) of each genotype. The DGE represents the "pure" performance of the genotype, while the IGE reflects the average effect of the same genotype on the genotypic value of its neighbours.

Competition classes

A higher IGE indicates a more aggressive genotype. In this study, we employ the classification proposed by Ferreira et al. (2023) to define competition classes.

$$\begin{cases} c_i > \overline{c} + sd(c) \times \tau \rightarrow \text{Aggressive} \\ \overline{c} + sd(c) \times \tau > c_i > \overline{c} - sd(c) \times \tau \rightarrow \text{Homeostatic} \\ c_i < \overline{c} - sd(c) \times \tau \rightarrow \text{Sensitive} \end{cases}$$

where \(c_i\) is the IGE of the ith genotype, \(\overline{c}\) is the mean IGE, \(sd(c)\) is the standard deviation of the IGE and \(\tau\) is a weight (defaults to 1).

Total genotypic value

The total genotypic value (TGV) of a candidate refers to its genetic merit summed to its competition capacity. When the competition effects are weighted by the distance between individuals - which applies to the methods adapted for forest data in gencomp, the TGV is given by:

$$TGV_i = d_i + CIF \times c_i$$

where \(d_i\) is the DGE of the ith genotype, and \(\phi\) is the overall competition intensity factor, previously computed in the function prepfor().

$$wTGV_i = r_{d_i}^2 \times d_i + r_{c_i}^2 \times {\phi \times c_i}$$

When there is no weighting based on distances (the case of agronomic data), the full contribution of IGE to the heritable variance is considered (Bijma, 2014). In this case, the TGV is given by:

$$TGV_i = d_i + c_i$$ id est, \(\phi=1\).

In both cases, when weight.tgv = TRUE, the DGE and IGE will be multiplied by their respective reliabilities (\(r_{d_i}^2\) and \(r_{c_i}^2\)) (Ferreira et al., 2024)

Heritability

When cor = TRUE in asr or asr_ma, resp estimates the DGE heritability and the total heritability, given by, respectively (Bijma et al., 2007):

$$H^2 = \frac{\sigma^2_g}{\sigma^2_p}$$

$$H^2_t = \frac{\sigma^2_t}{\sigma^2_p}$$

where \(\sigma^2_g\) is the DGE variance, \(\sigma^2_p\) is the phenotypic variance, and \(\sigma^2_t\) is the total heritable variance, given by \(\sigma^2_t = \sigma^2_g + 2 \times \phi \times \sigma_{gc} + \phi^2 \times \sigma^2_c\) for forestry data, and \(\sigma^2_t = \sigma^2_g + 2 \times \sigma_{gc} + \sigma^2_c\) for agronomic data. In these equations, \(\sigma_{gc}\) is the covariance between DGE and IGE estimated in the model, and \(\sigma^2_c\) is the IGE variance.

References

Bijma, P., Muir, W.M., Van Arendonk, J.A.M., 2007. Multilevel Selection 1: Quantitative genetics of inheritance and response to selection. Genetics 175, 277–288. doi:10.1534/genetics.106.062711

Bijma, P. 2014. The quantitative genetics of indirect genetic effects: a selective review of modelling issues. Heredity 112(1), 61-69. doi:10.1038/hdy.2013.15

Ferreira, F.M., Chaves, S.F.S., Bhering, L.L., Alves, R.S., Takahashi, E.K., Sousa, J.E., Resende, M.D.V, Leite, F.P., Gezan, S.A., Viana, J.M., Fernandes, S.B., Dias, K.O.G., 2023. A novel strategy to predict clonal composites by jointly modeling spatial variation and genetic competition. Forest Ecology and Management 548, 121393. doi:10.1016/j.foreco.2023.121393

Ferreira, F.M., Chaves, S.F.S., Santos, O.P., Nunes, A.C.P., Tambarussi, E.V., Pereira, G.S., Santos, G.A., Bhering, L.L., Dias, K.O.G., 2024. Competition effects can mislead selection in eucalypt breeding trials. Forest Ecology and Management 561, 121892. doi:10.1016/j.foreco.2024.121892

See also

Examples

# \donttest{
 library(gencomp)
 dat = euca[which(euca$age == "6y"),]
 comp_mat = prepfor(data = dat, gen = 'clone', area = 'area',
                   plt = 'tree', age = NULL, row = 'row', col = 'col',
                   dist.col = 3, dist.row = 2, trait = 'MAI', method = 'SK',
                   n.dec = 3, verbose = FALSE, effs = c("block"))
 model = asr(prep.out = comp_mat,
              fixed = MAI~ 1, 
              random = ~ block, 
              lrtest = TRUE, 
              spatial = TRUE, 
              cor = TRUE, 
              maxit = 20)
#> ASReml Version 4.2 21/01/2025 18:21:22
#>           LogLik        Sigma2     DF     wall
#>  1     -3053.104           1.0    786   18:21:22  (  2 restrained)
#>  2     -2981.157           1.0    786   18:21:22  (  1 restrained)
#>  3     -2886.687           1.0    786   18:21:22  (  1 restrained)
#>  4     -2834.395           1.0    786   18:21:22
#>  5     -2799.595           1.0    786   18:21:22
#>  6     -2783.301           1.0    786   18:21:22
#>  7     -2779.991           1.0    786   18:21:22  (  1 restrained)
#>  8     -2779.727           1.0    786   18:21:22
#>  9     -2779.723           1.0    786   18:21:23
#> Warning: Some components changed by more than 1% on the last iteration
#> ASReml Version 4.2 21/01/2025 18:21:23
#>           LogLik        Sigma2     DF     wall
#>  1     -2779.723           1.0    786   18:21:23
#>  2     -2779.723           1.0    786   18:21:23
#> gencomp: The model was updated and reached full convergence
#> ====> Starting likelihood ratio tests
#> ASReml Version 4.2 21/01/2025 18:21:23
#>           LogLik        Sigma2     DF     wall
#>  1     -2876.396           1.0    786   18:21:23
#>  2     -2844.537           1.0    786   18:21:23  (  1 restrained)
#>  3     -2822.942           1.0    786   18:21:23
#>  4     -2802.075           1.0    786   18:21:23
#>  5     -2791.460           1.0    786   18:21:23
#>  6     -2788.335           1.0    786   18:21:23
#>  7     -2788.116           1.0    786   18:21:23
#>  8     -2788.112           1.0    786   18:21:24
#> ASReml Version 4.2 21/01/2025 18:21:24
#>           LogLik        Sigma2     DF     wall
#>  1     -3007.350           1.0    786   18:21:24
#>  2     -2963.294           1.0    786   18:21:24
#>  3     -2921.223           1.0    786   18:21:24  (  4 restrained)
#>  4     -2904.930           1.0    786   18:21:24
#>  5     -2898.108           1.0    786   18:21:24
#>  6     -2896.700           1.0    786   18:21:24
#>  7     -2896.633           1.0    786   18:21:24
#>  8     -2896.628           1.0    786   18:21:24
#> ASReml Version 4.2 21/01/2025 18:21:24
#>           LogLik        Sigma2     DF     wall
#>  1     -2897.726           1.0    786   18:21:24
#>  2     -2861.826           1.0    786   18:21:24  (  1 restrained)
#>  3     -2836.455           1.0    786   18:21:24
#>  4     -2813.468           1.0    786   18:21:24
#>  5     -2801.403           1.0    786   18:21:24
#>  6     -2797.606           1.0    786   18:21:24
#>  7     -2797.275           1.0    786   18:21:24
#>  8     -2797.263           1.0    786   18:21:24
#>  9     -2797.262           1.0    786   18:21:24
#> ====> LRT results:
#>   effect LR-statistic    Pr(Chisq)
#> 1    DGE    217.03194 0.000000e+00
#> 2    IGE     18.29996 9.435551e-06
             
 results = resp(prep.out = comp_mat, model = model, weight.tgv = FALSE, sd.class = 1)
 # }